pyproject.toml
[project]
name = "imdb-sentiment-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.0",
    "joblib>=1.4.2",
    "python-dotenv>=1.0.1",
    "waitress>=3.0.2",
    "tensorflow>=2.12.0",
]
